+++ src/components/NotePreview.tsx
@@ -1,6 +1,5 @@
 import { ComponentProps } from "react";
-import { NoteInfo } from "@/models";
-import { cn, formatDateFromMs } from "@/libs";
+import { cn, formatDateFromMs, NoteInfo } from "@/libs";
 
 export type NotePreviewProps = NoteInfo & {
   isActive?: boolean;
