+++ src/libs/index.ts
@@ -1,4 +1,5 @@
 export * from "./constants";
+export * from "./models";
 
 export * from "./fs";
 export * from "./utils";
