+++ src/libs/models/note.ts
