+++ tsconfig.json
@@ -24,7 +24,6 @@
     "paths": {
       "@/components": ["src/components"],
       "@/hooks": ["src/hooks"],
-      "@/models": ["src/models"],
       "@/store": ["src/store"],
       "@/libs": ["src/libs"]
     }
