+++ vite.config.ts
@@ -16,7 +16,6 @@ export default defineConfig(async () => ({
     alias: {
       "@/components": resolve(__dirname, "src/components"),
       "@/hooks": resolve(__dirname, "src/hooks"),
-      "@/models": resolve(__dirname, "src/models"),
       "@/store": resolve(__dirname, "src/store"),
       "@/libs": resolve(__dirname, "src/libs"),
     },
